% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFV.R
\name{BFV_KeyGen}
\alias{BFV_KeyGen}
\title{Brakerski / Fan-Vercauteren}
\usage{
BFV_KeyGen(d = 4, q = 424242)
}
\arguments{
\item{d}{the d input}

\item{q}{the q input}
}
\value{
polynomial
}
\description{
Brakerski / Fan-Vercauteren
}
\examples{
d  =      4
n  =      2^d
p  =     (n/2)-1
q  = 424242
GenPolyMod(n)
}
