% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKKS.R
\name{compute_basis_coordinates}
\alias{compute_basis_coordinates}
\title{compute basis coordinates}
\usage{
compute_basis_coordinates(sigma_R_basis, z)
}
\arguments{
\item{sigma_R_basis}{sigma_R_basis}

\item{z}{z}
}
\value{
basis coordinates
}
\description{
Compute basis coordinates
}
