% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKKS.R
\name{coordinate_wise_random_rounding}
\alias{coordinate_wise_random_rounding}
\title{coordinate-wise random rounding}
\usage{
coordinate_wise_random_rounding(coordinates)
}
\arguments{
\item{coordinates}{coordinates}
}
\value{
rounded coordinates
}
\description{
Coordinate-wise random rounding
}
