% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{EncryptPoly0}
\alias{EncryptPoly0}
\title{Encrypt Polynomial Message Part 0}
\usage{
EncryptPoly0(m, pk0, u, e1, p, pm, q)
}
\arguments{
\item{m}{message}

\item{pk0}{public key part 0}

\item{u}{u}

\item{e1}{e1}

\item{p}{p}

\item{pm}{pm}

\item{q}{q}
}
\value{
polynomial which contains the message in ciphertext
}
\description{
Encrypt Polynomial Message Part 0
}
