% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{EncryptPoly1}
\alias{EncryptPoly1}
\title{Encrypt Polynomial Message Part 1}
\usage{
EncryptPoly1(pk1, u, e2, pm, q)
}
\arguments{
\item{pk1}{public key part 1}

\item{u}{u}

\item{e2}{e2}

\item{pm}{pm}

\item{q}{q}
}
\value{
polynomial which contains the message in ciphertext
}
\description{
Encrypt Polynomial Message Part 1
}
