% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FandV_CRT.R
\name{FandV_CRT}
\alias{FandV_CRT}
\title{Fan and Vercauteren encryption scheme with Chinese Remainder Theorem extension}
\usage{
p <- pars("FandV_CRT")
}
\description{
The Fan and Vercauteren scheme is implemented in this package, together with a
seamless implementation of modulus extension via the Chinese Remainder Theorem.
}
\details{
Description of the scheme.
}
\examples{
# Benchmark the performance of the scheme
#library(microbenchmark)
#p <- pars("FandV_CRT")
#microbenchmark({ keys <- keygen(p) }, unit="ms")
#microbenchmark({ ct1 <- enc(keys$pk, 2) }, unit="ms")
#ct2 <- enc(keys$pk, 3)
#microbenchmark({ ct1 + ct2 }, unit="ms")
#microbenchmark({ ct1 * ct2 }, unit="ms")
#microbenchmark({ dec(keys$sk, ct1) }, unit="ms")
}

