% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{GenError}
\alias{GenError}
\title{Generate a}
\usage{
GenError(n)
}
\arguments{
\item{n}{the order}
}
\value{
polynomial of order x^^n with discrete Gaussian distribution, bounded (not strictly true) by -n,n
}
\description{
Generate a
}
\examples{
n = 16
GenError(n)
}
