% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{GenEvalKey0}
\alias{GenEvalKey0}
\title{Generate the Evaluation Key}
\usage{
GenEvalKey0(a, s, e)
}
\arguments{
\item{a}{a}

\item{s}{s}

\item{e}{e}
}
\value{
polynomial
}
\description{
Generate the Evaluation Key
}
