% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\docType{import}
\name{GenPolyMod}
\alias{GenPolyMod}
\alias{CoefMod}
\title{Generate Polynomial Modulo}
\arguments{
\item{n}{the order}
}
\value{
polynomial of the form x^^n + 1
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{HEtools}{\code{\link[HEtools]{CoefMod}}}
}}

