% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{GenPubKey}
\alias{GenPubKey}
\title{Generate the Public Key}
\usage{
GenPubKey(a, n, e, pm)
}
\arguments{
\item{a}{a}

\item{n}{n}

\item{e}{e}

\item{pm}{pm}
}
\value{
list with the two polynomials that form the public key
}
\description{
Generate the Public Key
}
