% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{GenPubKey0}
\alias{GenPubKey0}
\title{Generate part 0 of the Public Key}
\usage{
GenPubKey0(a, s, e, pm, q)
}
\arguments{
\item{a}{a}

\item{s}{s}

\item{e}{e}

\item{pm}{pm}

\item{q}{q}
}
\value{
polynomial
}
\description{
Generate part 0 of the Public Key
}
