% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{GenSecretkey}
\alias{GenSecretkey}
\alias{GenSecretKey}
\title{Generate Secret key}
\usage{
GenSecretKey(n)
}
\arguments{
\item{n}{the order}
}
\value{
polynomial of order x^^n with coefficients (-1,-,1)
}
\description{
Generate Secret key
}
\examples{
n = 16
GenSecretKey(n)
}
