% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitives.R
\name{GenU}
\alias{GenU}
\title{Generate u}
\usage{
GenU(n)
}
\arguments{
\item{n}{the order}
}
\value{
polynomial of order x^^n-1 with coefficients (-1,-,1)
}
\description{
Generate u
}
\examples{
n = 16
GenU
}
