% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKKS.R
\name{round_coordinates}
\alias{round_coordinates}
\title{round coordinates}
\usage{
round_coordinates(coordinates)
}
\arguments{
\item{coordinates}{coordinates}
}
\value{
rounded coordinates
}
\description{
Round coordinates
}
