% Generated by roxygen2 (4.0.2): do not edit by hand
\name{saveFHE}
\alias{saveFHE}
\title{Save and load keys and ciphertexts}
\usage{
saveFHE(object, file)
}
\arguments{
\item{object}{the ciphertext or keys to be saved}

\item{file}{the filename to save (load) the object to (from)}
}
\description{
Use these functions rather than the base R functions for saving and loading
of ciphertexts and keys.
}
\details{
These functions provide a way to save ciphertexts and keys generated by this
package to a file in such a way that they can later be restored to another
R session (possibly on a different machine).

At present no compression is performed so for long-term storage or transmission
over a network it would be advisable to gzip the files after creation.
}
\examples{
p <- pars("FandV")
keys <- keygen(p)
ct1 <- enc(keys$pk, 2)
\dontrun{saveFHE(ct1, "~/myCipherText.fhe")}

# Start a new session or transfer to another machine
\dontrun{ct1 <- loadFHE("~/myCipherText.fhe")}
}

