% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKKS.R
\name{sigma_R_discretization}
\alias{sigma_R_discretization}
\title{sigma discretization}
\usage{
sigma_R_discretization(xi, M, z)
}
\arguments{
\item{xi}{xi}

\item{M}{M}

\item{z}{z}
}
\value{
sigma R discretization
}
\description{
Sigma discretization
}
